+++ benches/generators.rs
@@ -11,6 +11,8 @@
 extern crate test;
 extern crate rand;
 extern crate rand_isaac;
+extern crate rand_chacha;
+extern crate rand_hc128;
 extern crate rand_xorshift;
 
 const RAND_BENCH_N: u64 = 1000;
@@ -20,11 +22,11 @@ use std::mem::size_of;
 use test::{black_box, Bencher};
 
 use rand::prelude::*;
-use rand::prng::{Hc128Rng, ChaChaRng};
-use rand::prng::hc128::Hc128Core;
 use rand::rngs::adapter::ReseedingRng;
 use rand::rngs::{OsRng, JitterRng, EntropyRng};
 use rand_isaac::{IsaacRng, Isaac64Rng};
+use rand_chacha::ChaChaRng;
+use rand_hc128::{Hc128Rng, Hc128Core};
 use rand_xorshift::XorShiftRng;
 
 macro_rules! gen_bytes {