+++ rand_xorshift/CHANGELOG.md
@@ -5,4 +5,4 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [0.1.0] - 2018-07-16
-- Initial release
+- Pulled out of the Rand crate